+++ src/main/java/com/capstone/cschatbot/chat/service/communicate/GPTService.java
@@ -1,0 +1,7 @@
+package com.capstone.cschatbot.chat.service.communicate;
+
+import com.capstone.cschatbot.chat.entity.ChatRequest;
+
+public interface GPTService {
+    String getNewQuestion(ChatRequest chatRequest);
+}
